Share connect and disconnect menuitem through all android activities
