+++ client/android/res/menu/options_library_menu.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+	<item android:id="@+id/options_menu_connect" 
+		android:icon="@android:drawable/ic_menu_share"
+		android:title="@string/options_menu_connect"/>
+	<item android:id="@+id/options_menu_disconnect" android:title="@string/options_menu_disconnect"/>
+</menu>
