+++ client/android/src/remuco/client/android/Remuco.java
@@ -21,17 +21,12 @@
 
 package remuco.client.android;
 
-import remuco.client.android.io.WifiSocket;
-import remuco.client.android.dialogs.ConnectDialog;
 import remuco.client.android.dialogs.RatingDialog;
-import remuco.client.android.dialogs.VolumeDialog;
 import remuco.client.common.util.Log;
+
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -45,14 +40,6 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 public class Remuco extends RemucoActivity implements OnClickListener{
-
-	// --- dialog ids
-	private static final int CONNECT_DIALOG = 1;
-	private static final int VOLUME_DIALOG = 2;
-	private static final int RATING_DIALOG = 3;
-	
-	// --- dialog reference
-	private VolumeDialog volumeDialog;
 	
 	// --- view handler
 	private ViewHandler viewHandler;
@@ -166,29 +153,21 @@ public class Remuco extends RemucoActivity implements OnClickListener{
 	
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item) {
+        if (super.onOptionsItemSelected(item)) return true;
 		
 		switch(item.getItemId()){
 		
-		case R.id.options_menu_connect:
-			showDialog(CONNECT_DIALOG);
-			break;
-			
-		case R.id.options_menu_disconnect:
-			Log.ln("disconnect button pressed");
-			player.disconnect();
-			break;
-			
 		case R.id.options_menu_library:
             final Intent intent = new Intent(this, RemucoLibraryPlaylist.class);
             startActivity(intent);
-			break;
+			return true;
 			
 		case R.id.options_menu_rate:
 			showDialog(RATING_DIALOG);
-			break;
+			return true;
 		}
 		
-		return true;
+		return false;
 	}
 	
 	// ------------------------
@@ -197,85 +176,16 @@ public class Remuco extends RemucoActivity implements OnClickListener{
 	
 	@Override
 	protected Dialog onCreateDialog(int id) {
+        Dialog d = super.onCreateDialog(id);
+        if (d != null) return d;
 
 		switch(id){
-		
-		// --- connection dialog
-		case CONNECT_DIALOG:
-			
-			// create connect dialog
-			ConnectDialog cDialog = new ConnectDialog(this, player);
-
-			// register callback listener
-			
-			cDialog.setOnDismissListener(new DialogInterface.OnDismissListener() {
-				@Override
-				public void onDismiss(DialogInterface dialog) {
-                    player.disconnect();
-
-					// connect to host
-                    int type = ((ConnectDialog)dialog).getSelectedType();
-					String hostname = ((ConnectDialog)dialog).getSelectedHostname();
-                    int port = ((ConnectDialog)dialog).getSelectedPort();
-                    String bluedevice = ((ConnectDialog)dialog).getSelectedBluedevice();
-                    if (type == R.id.connect_dialog_wifi) {
-                        player.connectWifi(hostname, port, clientInfo);
-                    } else if (type == R.id.connect_dialog_bluetooth) {
-                        player.connectBluetooth(bluedevice, clientInfo);
-                    }
-					
-					// save new address in preferences
-					SharedPreferences.Editor editor = preference.edit();
-					editor.putInt(LAST_TYPE, type);
-					editor.putString(LAST_HOSTNAME, hostname);
-                    editor.putInt(LAST_PORT, port);
-                    editor.putString(LAST_BLUEDEVICE, bluedevice);
-					editor.commit();
-				}
-			});
-			
-			
-			return cDialog;
-			
-		// --- volume dialog
-		case VOLUME_DIALOG:
-			return new VolumeDialog(this, player);
-			
 		// --- rating dialog
 		case RATING_DIALOG:
 			return new RatingDialog(this, player);
 		}
-		
-		Log.bug("onCreateDialog(" + id + ") ... we shouldn't be here");
 		return null;
 	}
-	
-	@Override
-	protected void onPrepareDialog(int id, Dialog dialog) {
-		
-		switch(id){
-		
-		case CONNECT_DIALOG:
-			
-			ConnectDialog cDialog = (ConnectDialog)dialog;
-			
-			// set last hostname port
-			int type = preference.getInt(LAST_TYPE, R.id.connect_dialog_wifi);
-			cDialog.setType(type);
-			String hostname = preference.getString(LAST_HOSTNAME, "");
-			cDialog.setHostname(hostname);
-			int port = preference.getInt(LAST_PORT, WifiSocket.PORT_DEFAULT);
-			cDialog.setPort(port);
-            String bluedevice = preference.getString(LAST_BLUEDEVICE, "");
-			cDialog.setBluedevice(bluedevice);
-			
-			break;
-		
-		}
-		
-		
-	}
-	
 	
 	// --- handle clicks (this is the actual playback control)
 	
@@ -304,24 +214,5 @@ public class Remuco extends RemucoActivity implements OnClickListener{
 			
 		}
 	};	
-	
-	
-	
-	@Override
-	public boolean onKeyDown(int keyCode, KeyEvent event) {
-
-		switch(keyCode){
-		
-		case KeyEvent.KEYCODE_VOLUME_UP:
-			showDialog(VOLUME_DIALOG);
-			return true;
-			
-		case KeyEvent.KEYCODE_VOLUME_DOWN:
-			showDialog(VOLUME_DIALOG);
-			return true;
-		}
-		
-		return false;
-	}
 	
 }
